requirements.txt
torch
torchvision
onnx
onnxruntime
numpy
Pillow
fastapi
uvicorn[standard]
requests
pytest
